requirements.txt
flask
gradio_client
gunicorn
flask-cors
# google-generativeai
openai
python-dotenv